admin_feature

+++ handlers/employee.py
@@ -2,24 +2,24 @@ from datetime import datetime
 
 from aiogram import Router, F
 from aiogram.enums import ParseMode
-from aiogram.exceptions import TelegramNotFound
+from aiogram.exceptions import TelegramNotFound, TelegramBadRequest
 from aiogram.fsm.context import FSMContext
-from aiogram.types import Message, CallbackQuery
+from aiogram.types import Message, CallbackQuery, InlineKeyboardMarkup, InlineKeyboardButton
 from aiogram.filters import StateFilter, CommandStart
 from aiogram.utils.formatting import Bold, Text
 from babel.dates import get_month_names
 from dateutil.relativedelta import relativedelta
 
-from database.orm import get_employee, get_wh_statistics
+from database.orm import get_employee, get_wh_statistics, get_inquiries_by_employee_tab_no
 from handlers.fsm_states import Authorised, Unauthorised
 from handlers.utils import update_start_message, vsm_logo_uri
-from keyboards.inline import get_main_keyboard, get_start_keyboard, get_wh_info_keyboard
+from keyboards.inline import get_main_keyboard, get_start_keyboard, get_wh_info_keyboard, get_inquiry_menu_keyboard, \
+    get_back_button_keyboard
 
 employee_router = Router()
 
 @employee_router.message(StateFilter(Authorised), CommandStart())
 async def cmd_start(message: Message, state: FSMContext, session, _):
-    await message.delete()
 
     fsm_data = await state.get_data()
     tab_no = fsm_data.get('tab_no')
@@ -31,16 +31,22 @@ async def cmd_start(message: Message, state: FSMContext, session, _):
         '\u200B',
     )
 
-    start_msg_id = fsm_data.get('start_msg_id')
+    if message.text == '/start':
+        await message.delete()
 
-    if start_msg_id is not None:
-        try:
-            await message.bot.delete_message(chat_id=message.chat.id, message_id=start_msg_id)
-        except TelegramNotFound:
-            ...
+        start_msg_id = fsm_data.get('start_msg_id')
+        if start_msg_id is not None:
+            try:
+                await message.bot.delete_message(chat_id=message.chat.id, message_id=start_msg_id)
+            except TelegramBadRequest:
+                ...
 
-    start_msg = await message.answer_photo(photo=vsm_logo_uri, caption=welcome_message.as_markdown(), reply_markup=get_main_keyboard(_), parse_mode=ParseMode.MARKDOWN)
-    await state.update_data({'start_msg_id': start_msg.message_id})
+        start_msg = await message.answer_photo(photo=vsm_logo_uri, caption=welcome_message.as_markdown(), reply_markup=get_main_keyboard(_), parse_mode=ParseMode.MARKDOWN)
+        await state.update_data({'start_msg_id': start_msg.message_id})
+
+    else:
+        await update_start_message(message, state, welcome_message.as_markdown(), get_main_keyboard(_))
+
     await state.set_state(Authorised.start_menu)
 
 @employee_router.callback_query(StateFilter(Authorised.start_menu), (F.data == 'log_out_button'))
@@ -48,16 +54,20 @@ async def log_out(callback_query: CallbackQuery, state: FSMContext, _):
     await update_start_message(callback_query.message, state, '', get_start_keyboard(_))
     await state.set_state(Unauthorised.start_menu)
 
-@employee_router.callback_query(StateFilter(Authorised.wh_info), (F.data == 'back_button'))
+@employee_router.callback_query(
+    StateFilter(
+        Authorised.wh_info, Authorised.inquiry_menu,
+        Authorised.entering_inquiry_head, Authorised.entering_inquiry_body
+    ),
+    (F.data == 'back_button'))
 async def back(callback_query: CallbackQuery, state: FSMContext, session, _):
-    tab_no = (await state.get_data()).get('login_tab_no')
-    employee = await get_employee(session, tab_no)
-    welcome_message = Text(
-        '✅ ', Bold(_('Welcome'), ', ', ), '\n',
-        Bold('✅ ', employee.full_name),
-    )
-    await update_start_message(callback_query.message, state, welcome_message.as_markdown(), get_main_keyboard(_))
-    await state.set_state(Authorised.start_menu)
+    state_str = await state.get_state()
+    if state_str == 'Authorised:entering_inquiry_head':
+        await inquiry_menu(callback_query, state, session, _)
+    if state_str == 'Authorised:entering_inquiry_body':
+        await enter_inquiry_head(callback_query, state, session, _)
+    else:
+        await cmd_start(callback_query.message, state, session, _)
 
 @employee_router.callback_query(StateFilter(Authorised.start_menu), (F.data == 'get_wh_info'))
 async def get_wh_info(callback_query: CallbackQuery, state: FSMContext, session, _):
@@ -94,3 +104,43 @@ async def get_wh_info(callback_query: CallbackQuery, state: FSMContext, session,
 
     await update_start_message(callback_query.message, state, wh_info_message.as_markdown(), get_wh_info_keyboard(_))
     await state.set_state(Authorised.wh_info)
+
+@employee_router.callback_query(StateFilter(Authorised.start_menu), (F.data == 'inquiry_menu'))
+async def inquiry_menu(callback_query: CallbackQuery, state: FSMContext, session, _):
+    tab_no = (await state.get_data()).get('tab_no')
+    inquiries = await get_inquiries_by_employee_tab_no(session, tab_no)
+    inquiries_message = Text('🚹 ', _('Here you can write an inquiry regarding employment relations and workflow.'), '\n\n')
+    if not inquiries:
+        inquiries_message += Text(
+            ' 🔘 ', _('You have no inquiries yet.'), '\n'
+        )
+    else:
+        ...
+    await update_start_message(callback_query.message, state, inquiries_message.as_markdown(), get_inquiry_menu_keyboard(_))
+    state = await state.set_state(Authorised.inquiry_menu)
+
+@employee_router.callback_query(StateFilter(Authorised.inquiry_menu), (F.data == 'write_inquiry'))
+async def enter_inquiry_head(callback_query: CallbackQuery, state: FSMContext, session, _):
+    enter_inquiry_head_message = Text(
+        Bold(
+            '🚹 ',
+            _('Enter a title for your message. '
+              'It will be displayed in the list of requests, so it should be brief and meaningful '), ' ⤵️', '\n'
+        )
+    )
+    await update_start_message(callback_query.message, state, enter_inquiry_head_message.as_markdown(),
+                               get_back_button_keyboard(_))
+    await state.set_state(Authorised.entering_inquiry_head)
+
+@employee_router.message(StateFilter(Authorised.entering_inquiry_head))
+async def process_inquiry_head(message: Message, state: FSMContext, session, _):
+    await state.update_data({'inquiry_head': message.text})
+    enter_inquiry_body_message = Text(
+        Bold(
+            '🚹 ',
+            _('Enter the text of your inquiry'), ' ⤵️', '\n'
+        )
+    )
+    await update_start_message(message, state, enter_inquiry_body_message.as_markdown(), get_back_button_keyboard(_))
+    await state.set_state(Authorised.entering_inquiry_body)
+    ...
